notifications: Merge notificationsCommon.js into main file.
`notificationsCommon.js` contained the common code for notifications
and the `notifications.ios` and `notifications.android.js` contained
the platform-specific code. Now the two files belong together.
